feat: batch up electron-fiddle to send to broker
Right now the runner contacts the broker every single time there's
a line of log data to append from electron-fiddle.

This PR adds a debounce timer to batch them up and reduce the number of
round trips needed.

The current interval is 2 seconds. This is configurable via code so that
tests don't have to wait that long. I don't see a use case for end users
so I didn't add an environment variable, but I don't mind adding it if
a reviewer wants it.