fix: make broker's Server.close() async
Used to return void. Now return a Promise<void> that resolves when the
callback in net.Server.close(callback) is called.

This solves timing issues that I was hitting in the tests -- I think the
previous test's server wasn't done closing before the current test tried
to re-open on the same port.