build: add tsconfig.eslint.json to the submodules.
The dual-config approach you hit on is what upstream recommends too.
https://github.com/typescript-eslint/typescript-eslint/issues/1162

I've swapped 'build' to be the primary config and 'eslint' to be the
derived config since it makes for slightly simpler config: the diff in
the derived config is smaller, and only one other file (.eslintrc) is
affected instead of many (all the package.json files).