feat: run Test tasks in the runner (#115)
* feat: have runner bisect w/obsolete, beta, nightly

When running `electron-fiddle bisect`, include arguments to pull in all
of the Electron releases so that the bisection range will be as small as
possible.

Fixes #66 as a side-effect.

* feat: support + validate new test tasks in broker

Replaces some very bad broker validation code (of mine) with some new ow
predicates. This includes a TestJob predicate so the broker now supports
both Test and Bisect tasks.

Fixes #110.

* test: add test calling getJob() for an unknown job

* refactor: add JobType enum

* feat: support running individual tests in Runner.

Fixes #104.