build: import sibling workspaces from src/, not build/
The goal is to get cross-workspace code picked up in our test coverage.
This change pulls in the `shared/` workspace into our coverage reports.

This works; but since I didn't find any good docs on best practices for
monorepo imports like this, I'm not completely sure it's the preferred
way to do it. Were there other reasons that `build/` was used?